+++ app/commands/characters_context/dc20/feats/add_command.rb
@@ -13,17 +13,25 @@ module CharactersContext
 
         private
 
-        def do_persist(input) # rubocop: disable Metrics/AbcSize
+        def do_persist(input) # rubocop: disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/MethodLength, Metrics/PerceivedComplexity
           return { result: :ok } if ::Character::Feat.exists?(input)
 
           ActiveRecord::Base.transaction do
-            ::Character::Feat.create!(input)
+            ::Character::Feat.create!(input.merge(ready_to_use: true))
 
             input[:feat].info['rewrite']&.each { |key, value| input[:character].data[key] = value }
             input[:feat].info['increase']&.each { |key, value| input[:character].data[key] += value }
+            input[:feat].info['merge']&.each do |key, values|
+              attribute = input[:character].data[key]
+              values.each do |value_key, value|
+                attribute = attribute.merge({ attribute[value_key] => attribute[value_key] + value })
+              end
+            end
 
-            input[:character].data['classes'][input[:feat].origin_value] ||= 0
-            input[:character].data['classes'][input[:feat].origin_value] += 1
+            if input[:feat].origin == 'class'
+              input[:character].data['classes'][input[:feat].origin_value] ||= 0
+              input[:character].data['classes'][input[:feat].origin_value] += 1
+            end
 
             input[:character].save!
           end