+++ app/controllers/frontend/characters/feats_controller.rb
@@ -36,6 +36,7 @@ module Frontend
         when 'dnd5', 'dnd2024' then authorized_scope(Character.all).dnd
         when 'pathfinder2' then authorized_scope(Character.all).pathfinder2
         when 'daggerheart' then authorized_scope(Character.all).daggerheart
+        when 'dc20' then authorized_scope(Character.all).dc20
         else Character.none
         end
       end