+++ app/decorators/dc20_character/base_decorator.rb
@@ -3,10 +3,10 @@
 module Dc20Character
   class BaseDecorator < SimpleDelegator
     delegate :id, :name, :data, :feats, to: :__getobj__
-    delegate :abilities, :main_class, :level, :combat_expertise, :health, :classes, :attribute_points, :ancestries, :skill_levels,
+    delegate :abilities, :main_class, :level, :combat_expertise, :classes, :attribute_points, :ancestries, :skill_levels,
              :skill_expertise, :skill_points, :skill_expertise_points, :trade_points, :trade_expertise_points, :language_points,
              :trade_levels, :trade_expertise, :trade_knowledge, :language_levels, :conditions, :stamina_points, :path_points,
-             :paths, :mana_points, :maneuvers, :rest_points, :path, :spell_list, :talent_points, to: :data
+             :paths, :mana_points, :maneuvers, :rest_points, :path, :spell_list, :talent_points, :speeds, to: :data
 
     def parent = __getobj__
 
@@ -110,14 +110,22 @@ module Dc20Character
         .pick('items.info')
     end
 
+    def max_health = 0
     def maneuver_points = 0
     def technique_points = 0
     def cantrips = 0
     def spell_lists_amount = 0
     def spells = 0
+    def size = 'medium'
 
     def grit_points
       @grit_points ||= data.grit_points.merge('max' => modified_abilities['cha'] + 2)
+    end
+
+    def health
+      @health ||= __getobj__.data.health.merge(
+        'death_threshold' => 0 - modified_abilities['prime'] - combat_mastery
+      )
     end
 
     private