+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - rendering dice rolls for Pathfinder 2 characters
 - starting tooltips for Pathfinder characters
 - render used trait for attacks for DH characters
+- selecting ancestries for DC20 characters
 
 ### Fixed
 - rendering skills boosts for D&D characters